# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Cheesy Arena API
  version: 1.0.0-draft
  description: >-
    Versioned API introduced alongside the existing server-rendered interface.
    This document is expanded as each migration phase is implemented.
servers:
  - url: /api/v1
paths:
...
  /admin/match-play/commands:
    post:
      operationId: executeMatchPlayControlCommand
      summary: Execute a non-scoring match play field-control command
      description: Admin session, CSRF token and Idempotency-Key are required. Currently supports setAudienceDisplay, setAllianceStationDisplay, toggleBypass, signalVolunteers, signalReset, startTimeout, setTimeoutDisplay and setTestMatchName. Repeating the same key and command/value returns replayed=true without reapplying the mutation. Match start, abort, result post and match advancement are not yet exposed.
      security: [{sessionCookie: []}]
      parameters:
        - {in: header, name: X-CSRF-Token, required: true, schema: {type: string}}
        - {in: header, name: Idempotency-Key, required: true, schema: {type: string, maxLength: 128}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [command]
              properties:
                command: {type: string, enum: [setAudienceDisplay, setAllianceStationDisplay, toggleBypass, signalVolunteers, signalReset, startTimeout, setTimeoutDisplay, setTestMatchName]}
                value: {}
      responses:
        "200": {description: Command accepted or safely replayed}
        "400": {description: Malformed or unknown command}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {description: Idempotency key reused or command disallowed in current state}
        "422": {description: Invalid command value}
        "503": {description: Field command failed without leaking internal details}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Forbidden:
      description: Administrator permission or a valid CSRF token is required
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
  schemas:
...
    ResponseMeta:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: string
        version:
          type: integer
          minimum: 0
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
    ErrorResponse:
      type: object
      required:
        - error
        - meta
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...